Update Excel file generation for workbook and template download.
Made several modifications to enhance the functionality of the Excel workbook and template download features:

Workbook Download:

Added the capability to fetch records for each sheet.
Improved error handling to catch cases where records for a sheet might not be available.
Streamlined the process of writing data to the Excel workbook by iterating over the records fetched.
Template Download:

Reformatted the Excel template to have fields as rows and attributes as columns.
Introduced a new row to display the field key, followed by rows for each of the field attributes including Field Label, Description, Type, Required status, Unique status, and Readonly status.
Enhanced the handling of the field attributes to better reflect their status in the Excel template (e.g., displaying 'Yes' or 'No' for constraints).
These changes aim to provide a more intuitive and informative Excel download experience for the end-users, ensuring that they have all the necessary information in a structured format.